Move beta PDF computation to calculations.js
