+++ src/shared/calculations.js
@@ -1,5 +1,6 @@
 import quantile from "@stdlib/stats/base/dists/beta/quantile";
 import pdf from "@stdlib/stats/base/dists/beta/pdf";
+import linspace from "@stdlib/math/utils/linspace";
 
 /** @typedef ConfidenceInterval
  *  @property {number} proportion  proportion of ratings which were positive
@@ -116,6 +117,24 @@ export default {
         const proportion = this.getProportionPositive(ratingData, method);
         const numPositive = Math.floor(proportion * numRatings);
         return getBetaConfidenceInterval(numRatings, numPositive, 0.95);
+    },
+
+    /**
+     * Get [x, y] coordinates of a beta distribution.
+     * @param {number} alpha - "alpha" parameter of beta distribution
+     * @param {number} beta - "beta" parameter of beta distribution
+     * @param {number} resolution - number of steps to split the domain [0, 1] into
+     * @returns {number[][]} dist - array of [x, y] coordinate pairs
+     */
+    getBetaPDF(alpha, beta, resolution = 200) {
+        var dist = [];
+        const xs = linspace(0, 1, resolution);
+        for (var i = 0; i < xs.length; i++) {
+            const x = xs[i] * 1.0;
+            const y = pdf(x, alpha, beta);
+            dist.push([x, y]);
+        }
+        return dist;
     }
 }
